Give PendingActivations its own module
