+++ src/lib.rs
@@ -13,6 +13,7 @@ pub mod protos;
 
 pub(crate) mod core_tracing;
 mod machines;
+mod pending_activations;
 mod pollers;
 mod protosext;
 mod workflow;
@@ -25,6 +26,7 @@ pub use url::Url;
 
 use crate::{
     machines::{InconvertibleCommandError, WFCommand, WFMachinesError},
+    pending_activations::{PendingActivation, PendingActivations},
     protos::{
         coresdk::{
             task_completion, wf_activation_completion::Status, Task, TaskCompletion,
@@ -37,7 +39,6 @@ use crate::{
     protosext::HistoryInfoError,
     workflow::{NextWfActivation, WorkflowConcurrencyManager},
 };
-use crossbeam::queue::SegQueue;
 use dashmap::DashMap;
 use std::{
     convert::TryInto,
@@ -110,14 +111,6 @@ pub fn init(opts: CoreInitOptions) -> Result<impl Core> {
     })
 }
 
-/// Type of task queue to poll.
-pub enum TaskQueue {
-    /// Workflow task
-    Workflow(String),
-    /// Activity task
-    _Activity(String),
-}
-
 struct CoreSDK<WP>
 where
     WP: ServerGatewayApis + 'static,
@@ -136,46 +129,6 @@ where
 
     /// Has shutdown been called?
     shutdown_requested: AtomicBool,
-}
-
-/// Tracks pending activations using an internal queue, while also allowing fast lookup of any
-/// pending activations by run ID
-#[derive(Default)]
-struct PendingActivations {
-    queue: SegQueue<PendingActivation>,
-    count_by_id: DashMap<String, usize>,
-}
-
-impl PendingActivations {
-    pub fn push(&self, v: PendingActivation) {
-        *self
-            .count_by_id
-            .entry(v.run_id.clone())
-            .or_insert_with(|| 0)
-            .value_mut() += 1;
-        self.queue.push(v);
-    }
-
-    pub fn pop(&self) -> Option<PendingActivation> {
-        let rme = self.queue.pop();
-        if let Some(pa) = &rme {
-            if let Some(mut c) = self.count_by_id.get_mut(&pa.run_id) {
-                *c.value_mut() -= 1
-            }
-            self.count_by_id.remove_if(&pa.run_id, |_, v| v <= &0);
-        }
-        rme
-    }
-
-    pub fn has_pending(&self, run_id: &str) -> bool {
-        self.count_by_id.contains_key(run_id)
-    }
-}
-
-#[derive(Debug)]
-struct PendingActivation {
-    run_id: String,
-    task_token: Vec<u8>,
 }
 
 impl<WP> Core for CoreSDK<WP>
